requirements.txt
google-generativeai
aiomysql
python-dotenv
rich
openai
openai-agents
